Refactor: 탭 관련 상수 및 타입 분리
