+++ src/features/posts/constants/tabs.ts
@@ -1,0 +1,9 @@
+import { type TabType } from '../types/tabs';
+
+export const TAB_LABELS: Record<TabType, string> = {
+  recent: '최신글',
+  popular: '인기글',
+  my: '내 상담글',
+  commented: '댓글 단 글',
+  liked: '좋아한 글',
+};
