+++ src/features/posts/index.ts
@@ -1,6 +1,5 @@
 export { useCreatePost } from './ui/use-create-post';
 export { WriteButton } from './ui/write-button';
 export { fetchPostsByTab, fetchPostDetail } from './api/postApi';
-export type { TabType } from './api/postApi';
 export * from './api/mutations';
 export * from './actions/post-actions';
