+++ src/features/posts/types/tabs.ts
@@ -1,0 +1,1 @@
+export type TabType = 'recent' | 'popular' | 'my' | 'commented' | 'liked';
